Refactor: Move magic_metadata.dart under models/metadata
