+++ lib/models/metadata/magic_metadata.dart
