+++ lib/ui/viewer/file/file_details_widget.dart
@@ -5,7 +5,7 @@ import "package:photos/core/configuration.dart";
 import "package:photos/generated/l10n.dart";
 import "package:photos/models/file.dart";
 import "package:photos/models/file_type.dart";
-import "package:photos/models/magic_metadata.dart";
+import 'package:photos/models/metadata/magic_metadata.dart';
 import "package:photos/services/file_magic_service.dart";
 import "package:photos/services/update_service.dart";
 import 'package:photos/theme/ente_theme.dart';
@@ -177,7 +177,7 @@ class _FileDetailsWidgetState extends State<FileDetailsWidget> {
       ]);
     }
 
-    if(!UpdateService.instance.isFdroidFlavor()) {
+    if (!UpdateService.instance.isFdroidFlavor()) {
       fileDetailsTiles.addAll([
         ObjectsItemWidget(file),
         const FileDetailsDivider(),
