+++ lib/utils/trash_diff_fetcher.dart
@@ -3,7 +3,7 @@ import 'dart:math';
 
 import 'package:logging/logging.dart';
 import 'package:photos/core/network/network.dart';
-import 'package:photos/models/magic_metadata.dart';
+import 'package:photos/models/metadata/magic_metadata.dart';
 import 'package:photos/models/trash_file.dart';
 import 'package:photos/utils/crypto_util.dart';
 import 'package:photos/utils/file_download_util.dart';
